Copy parts of eadwu's gpu config.
This is based on [this
config](https://github.com/eadwu/nixos-configuration/tree/master/hardware) (or
at least as much as I could figure out), which is a configuration for
a Dell XPS 9570 (same system as mine) by a user who does not
experience tearing. However, this doesn't seem to be it.